chore: Remove CWE embedding and JSON logic from service
CWE Data is now pre-populated over at the DB
