+++ go.mod
@@ -6,7 +6,6 @@ require (
 	github.com/Ullaakut/nmap/v2 v2.2.2
 	github.com/google/uuid v1.6.0
 	github.com/kptm-tools/common v1.7.5
-	github.com/likexian/gokit v0.25.15
 	github.com/lmittmann/tint v1.1.2
 	github.com/nats-io/nats.go v1.43.0
 	github.com/stretchr/testify v1.10.0
@@ -17,6 +16,7 @@ require (
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/klauspost/compress v1.18.0 // indirect
+	github.com/likexian/gokit v0.25.15 // indirect
 	github.com/likexian/whois-parser v1.24.20 // indirect
 	github.com/nats-io/nkeys v0.4.11 // indirect
 	github.com/nats-io/nuid v1.0.1 // indirect
