--- pkg/services/cwe.go
@@ -1,182 +1,0 @@
-package services
-
-import (
-	"embed"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
-
-	"github.com/kptm-tools/common/common/pkg/results/tools"
-	"github.com/kptm-tools/vulnerability-analysis/pkg/dto"
-)
-
-//go:embed testdata/cwe.json
-var cweDataFS embed.FS
-
-var ErrCWENotFound = errors.New("CWE not found")
-
-// ICWEService represents operations on CWE data
-type ICWEService interface {
-	GetMitigationsFromWeakness(string) ([]tools.CWERemediation, error)
-}
-
-// CWEService holds a map of CWEWeakness indexed by CWE ID
-type CWEService struct {
-	Data map[string]dto.CWEWeakness
-}
-
-var (
-	cweServiceInstance *CWEService
-	cweInitOnce        sync.Once
-)
-
-// NewCWEService initializes CWEService by loading CWE data
-func NewCWEService() (*CWEService, error) {
-	var initErr error
-	cweInitOnce.Do(func() {
-		data, err := LoadCWE()
-		if err != nil {
-			initErr = err
-			return
-		}
-		cweServiceInstance = &CWEService{Data: data}
-	})
-	return cweServiceInstance, initErr
-}
-
-// LoadCWE reads the embedded JSON and builds a map of CWEWeakness by ID
-func LoadCWE() (map[string]dto.CWEWeakness, error) {
-	// Read file from embedded filesystem
-	raw, err := cweDataFS.ReadFile("testdata/cwe.json")
-	if err != nil {
-		return nil, fmt.Errorf("failed to read cwe.json: %w", err)
-	}
-
-	// Unmarshal JSON into CWEDetails struct
-	var all dto.CWEDetails
-	if err := json.Unmarshal(raw, &all); err != nil {
-		return nil, fmt.Errorf("error parsing JSON: %w", err)
-	}
-
-	// Build map: ID -> CWEWeakness
-	m := make(map[string]dto.CWEWeakness, len(all.Weaknesses))
-	for _, w := range all.Weaknesses {
-		m[w.ID] = w
-	}
-	return m, nil
-}
-
-// GetMitigationsFromWeakness gets the MITRE mitigations associated to a particular weakness.
-// It takes in a rawID (e.g., "CWE-749"), attempts to extract the numeric part and then
-// searches for mitigations.
-func (s *CWEService) GetMitigationsFromWeakness(rawID string) ([]tools.CWERemediation, error) {
-	cweID, err := parseWeaknessIDInt(rawID)
-	if err != nil {
-		var noInfo CWENoInfoError
-		var other CWEOtherError
-		if errors.As(err, &noInfo) || errors.As(err, &other) {
-			return []tools.CWERemediation{{ID: rawID}}, nil // No remediations for these two.
-		}
-		return nil, fmt.Errorf("failed to parse weakness ID: %w", err)
-	}
-
-	weakness, ok := s.Data[cweID]
-	if !ok {
-		return []tools.CWERemediation{}, fmt.Errorf("%w: %s", ErrCWENotFound, cweID)
-	}
-
-	// Check if the weakness has any mitigations
-	if len(weakness.PotentialMitigations) == 0 {
-		return []tools.CWERemediation{}, nil
-	}
-
-	// Find the most recent ModificationDate in weakness.ContentHistory
-	var lastUpdated time.Time
-	for _, h := range weakness.ContentHistory {
-		if h.ModificationDate == "" {
-			continue
-		}
-		dt, err := time.Parse("2006-01-02", h.ModificationDate)
-		if err != nil {
-			// skip malformed dates
-			continue
-		}
-		if dt.After(lastUpdated) {
-			lastUpdated = dt
-		}
-	}
-
-	// 4) convert dto.ProtentialMitigations -> tools.CWERemediation
-	out := make([]tools.CWERemediation, len(weakness.PotentialMitigations))
-	for i, m := range weakness.PotentialMitigations {
-		out[i] = tools.CWERemediation{
-			ID:                 rawID,
-			Title:              weakness.Name,
-			MitigationID:       m.MitigationID,
-			Phase:              m.Phase,
-			Description:        m.Description,
-			Effectiveness:      m.Effectiveness,
-			EffectivenessNotes: m.EffectivenessNotes,
-			LastUpdated:        lastUpdated,
-		}
-	}
-	return out, nil
-}
-
-// StandardizeCWEID converts raw CWE ID strings to standard "CWE-..." format.
-// Handles numeric-only strings (e.g., "928" -> "CWE-928") and special strings
-// (e.g., "NoInfo" -> "CWE-NoInfo"). Returns "CWE-noinfo" for empty/invalid input.
-func StandardizeCWEID(rawCWEID string) string {
-	// Handle empty or invalid input
-	if rawCWEID == "" {
-		return "CWE-noinfo"
-	}
-
-	// Already standardized
-	if strings.HasPrefix(rawCWEID, "CWE-") {
-		return rawCWEID
-	}
-
-	// Check if it's numeric only
-	if _, err := strconv.Atoi(rawCWEID); err == nil {
-		return "CWE-" + rawCWEID
-	}
-
-	// Handle special cases
-	switch strings.ToLower(rawCWEID) {
-	case "noinfo", "no info", "no-info":
-		return "CWE-NoInfo"
-	case "other":
-		return "CWE-Other"
-	default:
-		// For any other non-numeric string, prefix it as-is
-		return "CWE-" + rawCWEID
-	}
-}
-
-// parseWeaknessID extracts the integer part of a string like "CWE-79".
-// It also returns our custom CWENoInfoError or CWEOtherError when appropriate.
-func parseWeaknessIDInt(value string) (string, error) {
-	idx := strings.Index(value, "CWE-")
-	if idx < 0 {
-		return "", fmt.Errorf("invalid WeaknessID string: %s (missing 'CWE-')", value)
-	}
-	raw := value[idx+len("CWE-"):]
-	switch raw {
-	case "":
-		return "", fmt.Errorf("invalid WeaknessID string: %s (no number)", value)
-	case "noinfo":
-		return "", CWENoInfoError{}
-	case "Other":
-		return "", CWEOtherError{}
-	}
-	// At this point, raw should be all digits; we can (optionally) validate:
-	if _, err := strconv.Atoi(raw); err != nil {
-		return "", fmt.Errorf("invalid numeric portion in %q: %w", value, err)
-	}
-	return raw, nil
-}
