--- pkg/services/cwe_test.go
@@ -1,266 +1,0 @@
-package services_test
-
-import (
-	"testing"
-	"time"
-
-	"github.com/kptm-tools/vulnerability-analysis/pkg/dto"
-	"github.com/kptm-tools/vulnerability-analysis/pkg/services"
-	"github.com/likexian/gokit/assert"
-	"github.com/stretchr/testify/require"
-)
-
-// TestLoadCWE verifies that NewCWEService() loads data without error
-func TestLoadCWE(t *testing.T) {
-	svc, err := services.NewCWEService()
-	if err != nil {
-		t.Fatalf("expected no error initializing CWEService, got: %v", err)
-	}
-	if len(svc.Data) == 0 {
-		t.Fatal("expected at least one entry in svc.Data, but it's empty")
-	}
-}
-
-// TestMapIntegrity checks basic integrity of the weakness map
-func TestMapIntegrity(t *testing.T) {
-	svc, err := services.NewCWEService()
-	if err != nil {
-		t.Fatalf("error initializing CWEService: %v", err)
-	}
-
-	if len(svc.Data) == 0 {
-		t.Fatal("no weaknesses loaded into the map")
-	}
-
-	// Grab one example key/value
-	var exampleID string
-	exampleWeakness := struct {
-		ID   string
-		Name string
-	}{}
-
-	for k, w := range svc.Data {
-		exampleID = k
-		exampleWeakness.ID = w.ID
-		exampleWeakness.Name = w.Name
-		break
-	}
-
-	if exampleID == "" {
-		t.Error("example key from svc.Data is empty string")
-	}
-	if exampleWeakness.ID == "" {
-		t.Error("example CWEWeakness.ID is empty")
-	}
-	if exampleWeakness.Name == "" {
-		t.Error("example CWEWeakness.Name is empty")
-	}
-}
-
-func TestCWEService_GetMitigationsFromWeakness(t *testing.T) {
-	// Create a service instance with known test data.
-	testService := &services.CWEService{
-		Data: map[string]dto.CWEWeakness{
-			"79": {
-				ID:   "79",
-				Name: "Cross-site Scripting",
-				PotentialMitigations: []dto.CWEMitigation{
-					{MitigationID: "M-1", Phase: []string{"Implementation"}, Description: "Use output encoding."},
-					{MitigationID: "M-2", Phase: []string{"Architecture"}, Description: "Use Content Security Policy."},
-				},
-				ContentHistory: []dto.CWEContentHistory{
-					{ModificationDate: "2022-01-15"},
-					{ModificationDate: "2023-06-20"}, // Most recent
-					{ModificationDate: "2021-10-10"},
-				},
-			},
-			"89": {
-				ID:                   "89",
-				Name:                 "SQL Injection",
-				PotentialMitigations: []dto.CWEMitigation{}, // Valid CWE, but no mitigations listed
-				ContentHistory:       []dto.CWEContentHistory{},
-			},
-		},
-	}
-
-	t.Run("Happy Path: Should return all mitigations for a valid CWE", func(t *testing.T) {
-		// Act
-		remediations, err := testService.GetMitigationsFromWeakness("CWE-79")
-
-		// Assert
-		require.NoError(t, err)
-		require.NotNil(t, remediations)
-		require.Len(t, remediations, 2)
-
-		// Check content of the first remediation to ensure mapping is correct
-		assert.Equal(t, "CWE-79", remediations[0].ID)
-		assert.Equal(t, "Cross-site Scripting", remediations[0].Title)
-		assert.Equal(t, remediations[0].Phase, []string{"Implementation"})
-		assert.Equal(t, "Use output encoding.", remediations[0].Description)
-
-		// Check that LastUpdated was calculated correctly from the most recent history date
-		expectedDate, _ := time.Parse("2006-01-02", "2023-06-20")
-		assert.Equal(t, expectedDate, remediations[0].LastUpdated)
-	})
-
-	t.Run("Edge Case: Should handle CWE found but with no mitigations", func(t *testing.T) {
-		// Act
-		remediations, err := testService.GetMitigationsFromWeakness("CWE-89")
-
-		// Assert
-		require.NoError(t, err, "Expected no error, as this is a valid 'not found' case for mitigations")
-		require.NotNil(t, remediations)
-		require.Len(t, remediations, 0, "The function currently returns a slice with one item to signal this case.")
-		// assert.Equal(t, "CWE-89", remediations[0].ID, "The single item should only contain the ID")
-		// assert.Equal(t, "", remediations[0].Title)
-	})
-
-	t.Run("Failure Case: Should return an error when CWE is not in the data map", func(t *testing.T) {
-		// Act
-		remediations, err := testService.GetMitigationsFromWeakness("CWE-999") // This ID does not exist in our map
-
-		// Assert
-		require.Error(t, err)
-		assert.Len(t, remediations, 0, "The returned slice should be empty when an error occurs")
-	})
-
-	t.Run("Special Case: Should handle 'CWE-Other'", func(t *testing.T) {
-		// Act
-		remediations, err := testService.GetMitigationsFromWeakness("CWE-Other")
-
-		// Assert
-		require.NoError(t, err)
-		require.NotNil(t, remediations)
-		require.Len(t, remediations, 1)
-		assert.Equal(t, "CWE-Other", remediations[0].ID)
-		assert.Equal(t, "", remediations[0].Title)
-	})
-
-	t.Run("Failure Case: Should propagate error for invalid CWE format", func(t *testing.T) {
-		// Act
-		remediations, err := testService.GetMitigationsFromWeakness("NOT_A_VALID_CWE_ID")
-
-		// Assert
-		require.Error(t, err)
-		require.Nil(t, remediations, "On a critical error, the slice should be nil")
-		assert.Contains(t, err.Error(), "failed to parse weakness ID")
-	})
-}
-
-func TestStandardizeCWEID(t *testing.T) {
-	tests := []struct {
-		name     string
-		input    string
-		expected string
-	}{
-		// Already standardized cases
-		{
-			name:     "Already standardized numeric",
-			input:    "CWE-918",
-			expected: "CWE-918",
-		},
-		{
-			name:     "Already standardized Other",
-			input:    "CWE-Other",
-			expected: "CWE-Other",
-		},
-		{
-			name:     "Already standardized NoInfo",
-			input:    "CWE-NoInfo",
-			expected: "CWE-NoInfo",
-		},
-
-		// Numeric-only strings
-		{
-			name:     "Numeric only - single digit",
-			input:    "7",
-			expected: "CWE-7",
-		},
-		{
-			name:     "Numeric only - two digits",
-			input:    "79",
-			expected: "CWE-79",
-		},
-		{
-			name:     "Numeric only - three digits",
-			input:    "918",
-			expected: "CWE-918",
-		},
-		{
-			name:     "Numeric only - four digits",
-			input:    "1234",
-			expected: "CWE-1234",
-		},
-
-		// Special string cases
-		{
-			name:     "NoInfo",
-			input:    "NoInfo",
-			expected: "CWE-NoInfo",
-		},
-		{
-			name:     "noinfo lowercase",
-			input:    "noinfo",
-			expected: "CWE-NoInfo",
-		},
-		{
-			name:     "No Info with space",
-			input:    "No Info",
-			expected: "CWE-NoInfo",
-		},
-		{
-			name:     "no info lowercase with space",
-			input:    "no info",
-			expected: "CWE-NoInfo",
-		},
-		{
-			name:     "No-Info with hyphen",
-			input:    "No-Info",
-			expected: "CWE-NoInfo",
-		},
-		{
-			name:     "no-info lowercase with hyphen",
-			input:    "no-info",
-			expected: "CWE-NoInfo",
-		},
-		{
-			name:     "Other",
-			input:    "Other",
-			expected: "CWE-Other",
-		},
-		{
-			name:     "other lowercase",
-			input:    "other",
-			expected: "CWE-Other",
-		},
-
-		// Edge cases
-		{
-			name:     "Empty string",
-			input:    "",
-			expected: "CWE-noinfo",
-		},
-		{
-			name:     "Unknown string",
-			input:    "Unknown",
-			expected: "CWE-Unknown",
-		},
-		{
-			name:     "Custom string",
-			input:    "CustomVuln",
-			expected: "CWE-CustomVuln",
-		},
-		{
-			name:     "Mixed alphanumeric",
-			input:    "ABC123",
-			expected: "CWE-ABC123",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := services.StandardizeCWEID(tt.input)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
