+++ pkg/services/cwe_utils.go
@@ -1,0 +1,51 @@
+package services
+
+import (
+	"strconv"
+	"strings"
+)
+
+// CWENoInfoError represents when weakness ID is 'noinfo'
+type CWENoInfoError struct{}
+
+func (e CWENoInfoError) Error() string {
+	return "weakness ID is 'noinfo'"
+}
+
+// CWEOtherError represents when weakness ID is 'Other'
+type CWEOtherError struct{}
+
+func (e CWEOtherError) Error() string {
+	return "weakness ID is 'Other'"
+}
+
+// StandardizeCWEID converts raw CWE ID strings to standard "CWE-..." format.
+// Handles numeric-only strings (e.g., "928" -> "CWE-928") and special strings
+// (e.g., "NoInfo" -> "CWE-NoInfo"). Returns "CWE-noinfo" for empty/invalid input.
+func StandardizeCWEID(rawCWEID string) string {
+	// Handle empty or invalid input
+	if rawCWEID == "" {
+		return "CWE-noinfo"
+	}
+
+	// Already standardized
+	if strings.HasPrefix(rawCWEID, "CWE-") {
+		return rawCWEID
+	}
+
+	// Check if it's numeric only
+	if _, err := strconv.Atoi(rawCWEID); err == nil {
+		return "CWE-" + rawCWEID
+	}
+
+	// Handle special cases
+	switch strings.ToLower(rawCWEID) {
+	case "noinfo", "no info", "no-info":
+		return "CWE-NoInfo"
+	case "other":
+		return "CWE-Other"
+	default:
+		// For any other non-numeric string, prefix it as-is
+		return "CWE-" + rawCWEID
+	}
+}
