+++ pkg/services/nmap.go
@@ -16,19 +16,12 @@ import (
 )
 
 type NmapService struct {
-	cWEService ICWEService
 }
 
 var _ interfaces.INmapService = (*NmapService)(nil)
 
 func NewNmapService() *NmapService {
-	cweService, err := NewCWEService()
-	if err != nil {
-		panic(err)
-	}
-	return &NmapService{
-		cWEService: cweService,
-	}
+	return &NmapService{}
 }
 
 func (s *NmapService) RunScan(ctx context.Context, target string) (tools.ToolResult, error) {
@@ -295,7 +288,7 @@ func (s *NmapService) processNVDDataForPort(port nmap.Port, validCPE string) []t
 	for _, nvdVuln := range nvdData.Vulnerabilities {
 		var vuln tools.Vulnerability
 
-		if err := enrichVulnerabilityWithNvdData(&vuln, nvdVuln, s.cWEService); err != nil {
+		if err := enrichVulnerabilityWithNvdData(&vuln, nvdVuln); err != nil {
 			slog.Error("Failed to enrich vulnerability with nvd data, skipping to next vulnerability",
 				slog.String("host_name", port.Service.Hostname),
 				slog.Int("port_id", int(port.ID)),
@@ -363,7 +356,7 @@ func (s *NmapService) processNVDDataForOS(os tools.OSData) []tools.Vulnerability
 	for _, nvdVuln := range nvdData.Vulnerabilities {
 		var vuln tools.Vulnerability
 
-		if err := enrichVulnerabilityWithNvdData(&vuln, nvdVuln, s.cWEService); err != nil {
+		if err := enrichVulnerabilityWithNvdData(&vuln, nvdVuln); err != nil {
 			slog.Error("Failed to enrich OS vulnerability with nvd data, skipping to next vulnerability",
 				slog.String("os_name", os.Name),
 				slog.String("os_family", os.Family),
