+++ pkg/services/nvd.go
@@ -19,18 +19,6 @@ import (
 // sleeper is a function type that abstracts time.Sleep for testing
 type sleeper func(d time.Duration)
 
-type CWENoInfoError struct{}
-
-func (e CWENoInfoError) Error() string {
-	return "weakness ID is 'noinfo'"
-}
-
-type CWEOtherError struct{}
-
-func (e CWEOtherError) Error() string {
-	return "weakness ID is 'Other'"
-}
-
 var baseNvdAPIURL = "https://services.nvd.nist.gov/rest/json/cves/2.0"
 
 var ErrInvalidCPE = errors.New("invalid CPE name")
@@ -191,7 +179,6 @@ func standardizeCPE(cpe string) (string, error) {
 func enrichVulnerabilityWithNvdData(
 	vuln *tools.Vulnerability,
 	nvdVuln dto.Vulnerability,
-	cweService ICWEService,
 ) error {
 	if vuln == nil {
 		return fmt.Errorf("expected a non-nil vulnerability")
@@ -207,12 +194,8 @@ func enrichVulnerabilityWithNvdData(
 	vuln.CweID = StandardizeCWEID(rawID)
 	vuln.Type = enums.GetOwaspCategoryForCWE(rawID)
 
-	remediations, err := cweService.GetMitigationsFromWeakness(rawID)
-	if err != nil {
-		return fmt.Errorf("failed to get CWE mitigations: %w", err)
-	}
-
-	vuln.CWERemediation = remediations
+	// CWE remediations are now handled by the core-service from pre-populated database
+	vuln.CWERemediation = []tools.CWERemediation{}
 
 	// Descriptions - first english description
 	vuln.Description = getEnglishDescription(nvdVuln.Cve.Descriptions)
