+++ pkg/services/nvd_test.go
@@ -486,14 +486,7 @@ func Test_EnrichVulnerabilityWithNvdData(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			vuln := &tools.Vulnerability{} // Create a new vuln for each test
-			mockService := &CWEService{
-				Data: map[string]dto.CWEWeakness{
-					"744": {
-						ID: "744",
-					},
-				},
-			}
-			err := enrichVulnerabilityWithNvdData(vuln, tc.nvdVulnInput, mockService)
+			err := enrichVulnerabilityWithNvdData(vuln, tc.nvdVulnInput)
 
 			if tc.wantErr {
 				if err == nil {
@@ -757,85 +750,6 @@ func Test_parseVendorComments(t *testing.T) {
 				assert.Equal(t, wantComment.Comment, got[i].Comment)
 				assert.Equal(t, wantComment.LastModified, got[i].LastModified)
 			}
-		})
-	}
-}
-
-func Test_parseWeaknessIDInt(t *testing.T) {
-	testCases := []struct {
-		name             string // description of this test case
-		value            string
-		want             string
-		wantErr          bool
-		wantCWENoInfoErr bool
-		wantCWEOtherErr  bool
-	}{
-		{
-			name:             "Valid Weakness ID",
-			value:            "CWE-443",
-			want:             "443",
-			wantErr:          false,
-			wantCWENoInfoErr: false,
-			wantCWEOtherErr:  false,
-		},
-		{
-			name:             "No Info CWE",
-			value:            "CWE-noinfo",
-			want:             "",
-			wantErr:          true,
-			wantCWENoInfoErr: true,
-			wantCWEOtherErr:  false,
-		},
-		{
-			name:             "No Info CWE NVD",
-			value:            "NVD-CWE-noinfo",
-			want:             "",
-			wantErr:          true,
-			wantCWENoInfoErr: true,
-			wantCWEOtherErr:  false,
-		},
-		{
-			name:             "Other CWE",
-			value:            "NVD-CWE-Other",
-			want:             "",
-			wantErr:          true,
-			wantCWENoInfoErr: false,
-			wantCWEOtherErr:  true,
-		},
-		{
-			name:             "Other CWE",
-			value:            "CWE-Other",
-			want:             "",
-			wantErr:          true,
-			wantCWENoInfoErr: false,
-			wantCWEOtherErr:  true,
-		},
-		{
-			name:             "Empty String",
-			value:            "",
-			want:             "",
-			wantErr:          true,
-			wantCWENoInfoErr: false,
-			wantCWEOtherErr:  false,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			got, gotErr := parseWeaknessIDInt(tc.value)
-			if tc.wantErr {
-				assert.Error(t, gotErr)
-				if tc.wantCWENoInfoErr {
-					assert.ErrorAs(t, gotErr, &CWENoInfoError{}, "gotErr should be CWENoInfoError")
-				}
-				if tc.wantCWEOtherErr {
-					assert.NotNil(t, gotErr)
-					assert.ErrorAs(t, gotErr, &CWEOtherError{}, "gotErr should be CWEOtherError")
-				}
-				return
-			}
-
-			assert.NoError(t, gotErr, "gotErr should be false")
-			assert.Equal(t, got, tc.want, "parsed weakness ID should match want")
 		})
 	}
 }
